Address review: drop re-exports and side-effect hook
Apply reviewer feedback on the persistence cycle fixes:

- Delete the type re-exports from llm-usage-store.ts and
  llm-request-log-source.ts. Consumers now import the moved symbols directly
  from the leaf modules (usage-types.ts / llm-request-log-source-types.ts).
  The only external consumer of the moved usage symbols (usage-routes.ts) is
  updated accordingly.

- Replace the embedding retry-hook module (a module with load-time side
  effects) with a direct call. `startEmbeddingRuntimeManager` moves into
  embedding-backend.ts, where it calls `resetLocalEmbeddingFailureState`
  directly; embedding-runtime-manager.ts no longer imports the backend, so the
  cycle stays broken without any registration indirection. Deletes
  embedding-local-retry.ts and updates the lifecycle import.

- Rewrite the leaf module headers as present-tense contracts, dropping the
  refactor-history narration per the Code Comments convention.

madge circular chains unchanged at 415; still zero cycles touching the three
clusters.
