+++ assistant/src/persistence/embeddings/embedding-backend.ts
@@ -15,9 +15,9 @@ import {
   recordBillingSuccess,
 } from "./embedding-billing-breaker.js";
 import { GeminiEmbeddingBackend } from "./embedding-gemini.js";
-import { registerLocalEmbeddingRetryHook } from "./embedding-local-retry.js";
 import { OllamaEmbeddingBackend } from "./embedding-ollama.js";
 import { OpenAIEmbeddingBackend } from "./embedding-openai.js";
+import { EmbeddingRuntimeManager } from "./embedding-runtime-manager.js";
 import {
   EMBEDDING_DIMENSION_PROBE_TEXT,
   type EmbeddingBackend,
@@ -228,10 +228,34 @@ export function resetLocalEmbeddingFailureState(): void {
   }
 }
 
-// Wire the reset into the retry leaf so the embedding runtime manager can
-// clear sticky failure state after a background runtime download without
-// importing this module (which would form an import cycle).
-registerLocalEmbeddingRetryHook(resetLocalEmbeddingFailureState);
+/**
+ * Download the local embedding runtime in the background (non-blocking).
+ *
+ * Pre-warms the runtime (bun binary + ONNX worker scripts) so the first local
+ * embed doesn't pay the download cost inline. A no-op when the runtime is
+ * already installed. Failures are swallowed with a warning: local embeddings
+ * fall back to cloud backends, so a failed download must never block or crash
+ * daemon startup. Returns immediately; the download proceeds on its own.
+ */
+export function startEmbeddingRuntimeManager(): void {
+  void (async () => {
+    try {
+      const runtimeManager = new EmbeddingRuntimeManager();
+      if (runtimeManager.isReady()) return;
+      log.info("Downloading embedding runtime in background...");
+      await runtimeManager.ensureInstalled();
+      // Reset the sticky local-backend failure flag so auto mode retries local
+      // embeddings without evicting a worker that may already be live.
+      resetLocalEmbeddingFailureState();
+      log.info("Embedding runtime download complete");
+    } catch (err) {
+      log.warn(
+        { err },
+        "Embedding runtime download failed — local embeddings will use cloud fallback",
+      );
+    }
+  })();
+}
 
 function cacheKey(provider: string, model: string, extras?: string[]): string {
   if (extras && extras.length > 0) {
